requirements.txt
absl-py==0.12.0
aioredis==1.3.1
amqp==5.0.6
asgiref==3.4.1
astor==0.8.0
astunparse==1.6.3
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.2.1
Automat==20.2.0
billiard==3.6.4.0
cached-property==1.5.2
cachetools==4.2.1
celery==5.1.2
certifi==2020.12.5
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==3.0.4
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
construct==2.8.8
cryptography==36.0.0
cycler==0.10.0
daphne==3.0.2
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.9
django-crispy-forms==1.13.0
django-on-heroku==1.1.2
dm-tree==0.1.5
dronekit==2.9.2
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.27.1
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
#h5py==2.10.0
hexdump==3.3
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.7.3
incremental==21.3.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kombu==5.1.0
logutils==0.3.5
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.3.4
mock==4.0.3
monotonic==1.6
msgpack==1.0.2
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
pbr==5.5.1
Pillow==7.1.2
#pkg_resources==0.0.0
portpicker==1.3.1
prompt-toolkit==3.0.22
protobuf==3.15.6
psutil==5.8.0
psycopg2-binary==2.9.2
#ptpy ==./tx2/djenv/lib/python3.6/site-packages/ptpy
py-cpuinfo==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pymavlink==2.4.15
pyOpenSSL==21.0.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.3
pyusb==1.1.1
PyYAML==5.4.1
rainbow-logging-handler==2.2.2
redis==4.0.2
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
service-identity==21.1.0
six==1.15.0
sqlparse==0.4.2
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
#tensorflow @ file:///home/jetsontx2/Downloads/tensorflow-2.4.0%2Bnv21.4-cp36-cp36m-linux_aarch64.whl
tensorflow-estimator==2.4.0
tensorflow-model-optimization==0.7.0
termcolor==1.1.0
testresources==2.0.1
Twisted==21.7.0
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.25.11
vine==5.0.0
watchdog==2.1.6
wcwidth==0.2.5
Werkzeug==1.0.1
whitenoise==5.3.0
wrapt==1.12.1
zipp==3.4.1
zope.interface==5.4.0
